+++ src/device/main_windows.cpp
@@ -987,6 +987,10 @@ bool next_event(OsEvent &ev)
 
 struct InitGlobals
 {
+#if !CROWN_DEBUG && !CROWN_DEVELOPMENT
+	FreeConsole();
+#endif
+
 	InitGlobals()
 	{
 		memory_globals::init();
@@ -1007,21 +1011,6 @@ struct InitGlobals
 int main(int argc, char **argv)
 {
 	using namespace crown;
-
-	if (AttachConsole(ATTACH_PARENT_PROCESS) != 0) {
-		const DWORD handles[] = { STD_OUTPUT_HANDLE, STD_ERROR_HANDLE, STD_INPUT_HANDLE };
-		const char *modes[] = { "w", "w", "r" };
-		FILE *stdfds[] = { stdout, stderr, stdin };
-
-		for (u32 i = 0; i < countof(handles); ++i) {
-			HANDLE out = GetStdHandle(handles[i]);
-			int fd = _open_osfhandle((intptr_t)out, _O_TEXT);
-			if (fd != -1) {
-				*stdfds[i] = *_fdopen(fd, modes[i]);
-				setvbuf(stdfds[i], NULL, _IONBF, 0); // No buffering.
-			}
-		}
-	}
 
 	// code-format off
 	PHANDLER_ROUTINE signal_handler = [](DWORD dwCtrlType) {